examples(modelinversion): update model inversion example to use custom plugins
