+++ examples/tensorflow-mnist-model-inversion/docker-compose.yml
@@ -213,7 +213,7 @@ services:
     networks:
       - lab_deployment_backend
     volumes:
-      - nfs-datasets:/nfs/data/Mnist
+      - nfs-datasets:/nfs/data
       - restapi-data:/work/data
 
   tfcpu-02:
@@ -244,7 +244,7 @@ services:
     networks:
       - lab_deployment_backend
     volumes:
-      - nfs-datasets:/nfs/data/Mnist
+      - nfs-datasets:/nfs/data
       - restapi-data:/work/data
 
 volumes:
